requirements.txt
browser-use==0.12.6
playwright
mcp[server]
python-dotenv